+++ A97227/disk_usage.diff
@@ -1,0 +1,26 @@
+--- disk_usage_original.py	2024-01-31 16:11:53.018861500 +0300
++++ disk_usage_fixed.py	2024-01-31 16:12:54.616698000 +0300
+@@ -1,7 +1,7 @@
+ #!/usr/bin/env python3
+ 
+ import shutil
+-
++import sys
+ def check_disk_usage(disk, min_absolute, min_percent):
+     """Returns True if there is enough free disk space, false otherwise."""
+     du = shutil.disk_usage(disk)
+@@ -14,9 +14,9 @@
+     return True
+ 
+ # Check for at least 2 GB and 10% free
+-if not check_disk_usage("/", 2*2**30, 10):
++if not check_disk_usage("/", 2, 10):
+     print("ERROR: Not enough disk space")
+-    return 1
++    sys.exit(1)
+ 
+ print("Everything ok")
+-return 0
+\ No newline at end of file
++sys.exit(0)
+\ No newline at end of file
